% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{hfsis-package}
\alias{hfsis}
\alias{hfsis-package}
\title{hfsis: Adaptive spot support screening}
\description{
Implements M-SIS and PR-SISIS for equally spaced high-frequency regression
increments using coordinatewise truncation and pathwise active-set window
selection.
}
\references{
Zhu, H. \emph{Ultra-high-dimensional spot support screening in
continuous-time regression}. Manuscript.
}
\author{
\strong{Maintainer}: Haibin Zhu \email{haibinzhu@jnu.edu.cn}

}
\keyword{internal}
