% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{plot_cluster_share}
\alias{plot_cluster_share}
\title{Plot descriptive cluster selection shares}
\usage{
plot_cluster_share(x, ...)
}
\arguments{
\item{x}{A result from \code{\link[=cluster_share]{cluster_share()}}.}

\item{...}{Additional arguments passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
x, invisibly.
}
\description{
Plot descriptive cluster selection shares
}
